Rename quick_plot.py to standard_plots.py
